SmallFix

+++ resp/commodity.go
@@ -4,7 +4,7 @@ type Commodity struct {
 	Hash     string  `json:"hash"`
 	Image    string  `json:"image"`
 	Name     string  `json:"name"`
-	Price    float32 `json:"price"`
+	Price    float64 `json:"price"`
 	Amount   int     `json:"amount"`
 	OrderNum string  `json:"orderNum"`
 }
